moved the github APIs where they should be
