+++ treeherder/etl/bugzilla.py
@@ -3,7 +3,7 @@ import logging
 import dateutil.parser
 from django.conf import settings
 
-from treeherder.etl.common import fetch_json
+from treeherder.utils.github import fetch_json
 from treeherder.model.models import Bugscache
 
 logger = logging.getLogger(__name__)
