+++ treeherder/etl/pushlog.py
@@ -5,7 +5,7 @@ import newrelic.agent
 import requests
 from django.core.cache import cache
 
-from treeherder.etl.common import fetch_json
+from treeherder.utils.github import fetch_json
 from treeherder.etl.exceptions import CollectionNotStoredException
 from treeherder.etl.push import store_push
 from treeherder.model.models import Repository
