+++ treeherder/log_parser/artifactbuildercollection.py
@@ -2,7 +2,7 @@ import logging
 
 import newrelic.agent
 
-from treeherder.etl.common import make_request
+from treeherder.utils.github import make_request
 
 from .artifactbuilders import (BuildbotJobArtifactBuilder,
                                BuildbotLogViewArtifactBuilder,
