+++ treeherder/push_health/compare.py
@@ -1,6 +1,6 @@
 import logging
 
-from treeherder.etl.common import fetch_json
+from treeherder.utils.github import fetch_json
 from treeherder.model.models import (Commit,
                                      Push)
 from treeherder.webapp.api.serializers import RepositorySerializer
