+++ treeherder/services/pulse/consumers.py
@@ -8,7 +8,7 @@ from kombu import (Connection,
                    Queue)
 from kombu.mixins import ConsumerMixin
 
-from treeherder.etl.common import fetch_json
+from treeherder.utils.github import fetch_json
 from treeherder.etl.tasks.pulse_tasks import (store_pulse_pushes,
                                               store_pulse_tasks)
